# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockbusters Game Show API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Question {
  id: number;
  questionId: string;
  text: string;
  difficulty: string;
  category: string;
  /** @nullable */
  answer?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateQuestionInput {
  questionId: string;
  text: string;
  difficulty: string;
  category: string;
  /** @nullable */
  answer?: string | null;
}

export interface Tournament {
  id: number;
  name: string;
  event: string;
  division: string;
  createdAt: string;
}

export interface CreateTournamentInput {
  name: string;
  event: string;
  division: string;
}

export interface Match {
  id: number;
  matchNumber: number;
  /** @nullable */
  tournamentId?: number | null;
  redTeamName: string;
  blueTeamName: string;
  boardSize: string;
  status: string;
  /** @nullable */
  round?: string | null;
  /** @nullable */
  winnerTeam?: string | null;
  createdAt: string;
}

export interface CreateMatchInput {
  matchNumber: number;
  /** @nullable */
  tournamentId?: number | null;
  redTeamName: string;
  blueTeamName: string;
  boardSize: string;
  /** @nullable */
  round?: string | null;
}

export interface UpdateMatchInput {
  status?: string;
  /** @nullable */
  winnerTeam?: string | null;
  redTeamName?: string;
  blueTeamName?: string;
}

/**
 * @nullable
 */
export type GameStateCurrentQuestion = { [key: string]: unknown } | null;

export interface BlockState {
  index: number;
  /** @nullable */
  owner?: string | null;
  /** @nullable */
  questionId?: number | null;
  isMatchPoint: boolean;
}

export interface GameState {
  matchId: number;
  blocks: BlockState[];
  boardSize: string;
  redTeamName: string;
  blueTeamName: string;
  status: string;
  /** @nullable */
  winnerTeam?: string | null;
  /** @nullable */
  buzzerStatus?: string | null;
  /** @nullable */
  currentBlockIndex?: number | null;
  /** @nullable */
  currentQuestion?: GameStateCurrentQuestion;
}

export interface AwardBlockInput {
  blockIndex: number;
  team: string;
  /** @nullable */
  questionId?: number | null;
}

export interface SheetsSyncInput {
  spreadsheetId: string;
  sheetName?: string;
}

export interface SheetsSyncResult {
  synced: number;
  skipped: number;
  errors: number;
  message: string;
}

export type GetQuestionsParams = {
  category?: string;
  difficulty?: string;
};

export type GetMatchesParams = {
  /**
   * @nullable
   */
  tournamentId?: number | null;
  status?: string;
};