# lib/api-zod/src/generated/types/awardBlockInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockbusters Game Show API
 * OpenAPI spec version: 0.1.0
 */

export interface AwardBlockInput {
  blockIndex: number;
  team: string;
  /** @nullable */
  questionId?: number | null;
}