# lib/api-zod/src/generated/types/blockState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockbusters Game Show API
 * OpenAPI spec version: 0.1.0
 */

export interface BlockState {
  index: number;
  /** @nullable */
  owner?: string | null;
  /** @nullable */
  questionId?: number | null;
  isMatchPoint: boolean;
}