# lib/api-zod/src/generated/types/createMatchInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockbusters Game Show API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMatchInput {
  matchNumber: number;
  /** @nullable */
  tournamentId?: number | null;
  redTeamName: string;
  blueTeamName: string;
  boardSize: string;
  /** @nullable */
  round?: string | null;
}