# lib/api-zod/src/generated/types/createQuestionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockbusters Game Show API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQuestionInput {
  questionId: string;
  text: string;
  difficulty: string;
  category: string;
  /** @nullable */
  answer?: string | null;
}