# lib/api-zod/src/generated/types/createTournamentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockbusters Game Show API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTournamentInput {
  name: string;
  event: string;
  division: string;
}