# lib/api-zod/src/generated/types/gameState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockbusters Game Show API
 * OpenAPI spec version: 0.1.0
 */
import type { BlockState } from "./blockState";
import type { GameStateCurrentQuestion } from "./gameStateCurrentQuestion";

export interface GameState {
  matchId: number;
  blocks: BlockState[];
  boardSize: string;
  redTeamName: string;
  blueTeamName: string;
  status: string;
  /** @nullable */
  winnerTeam?: string | null;
  /** @nullable */
  buzzerStatus?: string | null;
  /** @nullable */
  currentBlockIndex?: number | null;
  /** @nullable */
  currentQuestion?: GameStateCurrentQuestion;
}