# lib/api-zod/src/generated/types/sheetsSyncInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockbusters Game Show API
 * OpenAPI spec version: 0.1.0
 */

export interface SheetsSyncInput {
  spreadsheetId: string;
  sheetName?: string;
}