# lib/api-zod/src/generated/types/tournament.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockbusters Game Show API
 * OpenAPI spec version: 0.1.0
 */

export interface Tournament {
  id: number;
  name: string;
  event: string;
  division: string;
  createdAt: string;
}